+++ refsanalysis.py
@@ -62,7 +62,7 @@ from forefst import (
 )
 
 PROG = "refsanalysis"
-VERSION = "1.10.2"
+VERSION = "1.10.3"